requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
pillow>=10.0
numpy>=1.24
tensorflow>=2.13, <2.18
python-multipart>=0.0.9
pandas>=2.0
matplotlib>=3.7
scikit-learn>=1.3